# backend/src/zod/types/complaint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyberpunk Pothole Detection & Routing System API
 * OpenAPI spec version: 0.1.0
 */

export interface Complaint {
  id: string;
  pothole_id: string;
  type: string;
  road_type: string;
  authority: string;
  description: string;
  contact_name?: string | null;
  contact_email?: string | null;
  status: string;
  created_at: Date;
}