# backend/src/zod/types/createPotholeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyberpunk Pothole Detection & Routing System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePotholeBodySeverity } from "./createPotholeBodySeverity";

export interface CreatePotholeBody {
  lat: number;
  lon: number;
  severity: CreatePotholeBodySeverity;
  depth_cm: number;
  width_cm: number;
  volume_m3: number;
  estimated_repair_cost_usd: number;
  image_base64?: string | null;
  confidence: number;
  address?: string | null;
}