# backend/src/zod/types/driverScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyberpunk Pothole Detection & Routing System API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverScore {
  score: number;
  level: string;
  potholes_reported: number;
  confirmations_received: number;
  city_savings_usd: number;
  badge: string;
  next_level_threshold: number;
}